+++ fuel-vm/src/interpreter/balances.rs
@@ -164,7 +164,7 @@ impl RuntimeBalances {
     where
         Tx: ExecutableTransaction,
     {
-        let len = vm.max_inputs() * (AssetId::LEN + WORD_SIZE) as Word;
+        let len = (vm.max_inputs() as usize * (AssetId::LEN + WORD_SIZE)) as Word;
 
         vm.registers[RegId::SP] += len;
         vm.reserve_stack(len).expect(
